+++ tests/Pgvector.CSharp.Tests/EntityFrameworkCoreTests.cs
@@ -2,6 +2,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Storage;
 using Pgvector.EntityFrameworkCore;
+using System.Collections;
 using System.ComponentModel.DataAnnotations.Schema;
 
 namespace Pgvector.Tests;
@@ -40,6 +41,9 @@ public class Item
     [Column("half_embedding", TypeName = "halfvec(3)")]
     public HalfVector? HalfEmbedding { get; set; }
 
+    [Column("binary_embedding", TypeName = "bit(3)")]
+    public BitArray? BinaryEmbedding { get; set; }
+
     [Column("sparse_embedding", TypeName = "sparsevec(3)")]
     public SparseVector? SparseEmbedding { get; set; }
 }
@@ -55,9 +59,9 @@ public class EntityFrameworkCoreTests
         var databaseCreator = ctx.GetService<IRelationalDatabaseCreator>();
         databaseCreator.CreateTables();
 
-        ctx.Items.Add(new Item { Embedding = new Vector(new float[] { 1, 1, 1 }), HalfEmbedding = new HalfVector(new Half[] { (Half)1, (Half)1, (Half)1 }), SparseEmbedding = new SparseVector(new float[] { 1, 1, 1 }) });
-        ctx.Items.Add(new Item { Embedding = new Vector(new float[] { 2, 2, 2 }), HalfEmbedding = new HalfVector(new Half[] { (Half)2, (Half)2, (Half)2 }), SparseEmbedding = new SparseVector(new float[] { 2, 2, 2 }) });
-        ctx.Items.Add(new Item { Embedding = new Vector(new float[] { 1, 1, 2 }), HalfEmbedding = new HalfVector(new Half[] { (Half)1, (Half)1, (Half)2 }), SparseEmbedding = new SparseVector(new float[] { 1, 1, 2 }) });
+        ctx.Items.Add(new Item { Embedding = new Vector(new float[] { 1, 1, 1 }), HalfEmbedding = new HalfVector(new Half[] { (Half)1, (Half)1, (Half)1 }), BinaryEmbedding = new BitArray(new bool[] { false, false, false }), SparseEmbedding = new SparseVector(new float[] { 1, 1, 1 }) });
+        ctx.Items.Add(new Item { Embedding = new Vector(new float[] { 2, 2, 2 }), HalfEmbedding = new HalfVector(new Half[] { (Half)2, (Half)2, (Half)2 }), BinaryEmbedding = new BitArray(new bool[] { true, false, true }), SparseEmbedding = new SparseVector(new float[] { 2, 2, 2 }) });
+        ctx.Items.Add(new Item { Embedding = new Vector(new float[] { 1, 1, 2 }), HalfEmbedding = new HalfVector(new Half[] { (Half)1, (Half)1, (Half)2 }), BinaryEmbedding = new BitArray(new bool[] { true, true, true }), SparseEmbedding = new SparseVector(new float[] { 1, 1, 2 }) });
         ctx.SaveChanges();
 
         var embedding = new Vector(new float[] { 1, 1, 1 });
@@ -86,6 +90,13 @@ public class EntityFrameworkCoreTests
         var sparseEmbedding = new SparseVector(new float[] { 1, 1, 1 });
         items = await ctx.Items.OrderBy(x => x.SparseEmbedding!.L2Distance(sparseEmbedding)).Take(5).ToListAsync();
         Assert.Equal(new int[] { 1, 3, 2 }, items.Select(v => v.Id).ToArray());
+
+        var binaryEmbedding = new BitArray(new bool[] { true, false, true });
+        items = await ctx.Items.OrderBy(x => x.BinaryEmbedding!.HammingDistance(binaryEmbedding)).Take(5).ToListAsync();
+        Assert.Equal(new int[] { 2, 3, 1 }, items.Select(v => v.Id).ToArray());
+
+        items = await ctx.Items.OrderBy(x => x.BinaryEmbedding!.JaccardDistance(binaryEmbedding)).Take(5).ToListAsync();
+        Assert.Equal(new int[] { 2, 3, 1 }, items.Select(v => v.Id).ToArray());
 
         items = await ctx.Items
             .OrderBy(x => x.Id)